split smoke tests and basic tests
