+++ tox.ini
@@ -177,37 +177,37 @@ basepython = python2.6
 install_command = pip install -U --force-reinstall {opts} {packages}
 deps = -rpy26-unit-tests-requirements.txt
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test/smoke_tests.py -s -vv --durations=5
 
 [testenv:py2.7-smoke-tests]
 passenv =
     SCALYR_API_KEY READ_API_KEY SCALYR_SERVER AGENT_HOST_NAME
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test -s -vv --durations=5
 
 [testenv:py3.5-smoke-tests]
 passenv =
     SCALYR_API_KEY READ_API_KEY SCALYR_SERVER AGENT_HOST_NAME
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test -s -vv --durations=5
 
 [testenv:py3.6-smoke-tests]
 passenv =
     SCALYR_API_KEY READ_API_KEY SCALYR_SERVER AGENT_HOST_NAME
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test -s -vv --durations=5
 
 [testenv:py3.7-smoke-tests]
 passenv =
     SCALYR_API_KEY READ_API_KEY SCALYR_SERVER AGENT_HOST_NAME
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test -s -vv --durations=5
 
 [testenv:py3.8-smoke-tests]
 passenv =
     SCALYR_API_KEY READ_API_KEY SCALYR_SERVER AGENT_HOST_NAME
 commands =
-    py.test tests/smoke_tests/standalone_test.py -s -vv --durations=5
+    py.test tests/smoke_tests/standalone_test -s -vv --durations=5
 
 [testenv:py3.5-smoke-tests-rate-limit]
 setenv =
